% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popstats.R
\name{pop_sd}
\alias{pop_sd}
\title{Population Standard Deviation}
\usage{
pop_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix, or data frame.}

\item{na.rm}{a logical: should \code{NA} values be removed?}
}
\value{
The population standard deviation of \code{x}
}
\description{
Population Standard Deviation
}
\examples{
pop_sd(1:10)
2.872281

pop_sd(c(1:10, NA))
NA

pop_sd(c(1:10, NA), na.rm = TRUE)
2.872281

}
