% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_var.R
\name{pop_var}
\alias{pop_var}
\title{Population Variance}
\usage{
pop_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix, or data frame.}

\item{na.rm}{a logical: should \code{NA} values be removed?}
}
\value{
The population variance of \code{x}
}
\description{
Population Variance
}
\examples{
pop_var(1:10)
8.25

pop_var(c(1:10, NA))
NA

pop_var(c(1:10, NA), na.rm = TRUE)
8.25

}
